refactor(unity-exec): extract plan and call pass executors
Why:
- OperationPhaseExecutor combined orchestration with per-phase loops and state tracking, which reduced cohesion and made execution-order changes hard to isolate.

What Changed:
- Added dedicated plan-pass and call-pass executors, result models, and shared phase utilities under Execution/Phases.
- Reduced OperationPhaseExecutor to orchestration: run plan pass, validate plan token, then run call pass.
- Preserved existing fail-fast and skipped-trace behavior while moving phase logic into focused classes.

Impact:
- Execution flow remains equivalent, but phase-specific behavior is now encapsulated and easier to review and test independently.
