+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/Phases/CallPassResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca2c1afb93c404a298cbb1927e1b3ee8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
