+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/Phases/IOperationCallPassExecutor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd790c744369844c3855a82affc038b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
