+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/Phases/PlanPassResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f05fc07ff2e5e4837997aec8dbb5f356
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
