+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/Phases/PreparedOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f3a184b888054ec890c28fa2bb10f86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
